+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.7 - 2023/mm/dd
 
 - ADD timestamp comparison on signup form submission against bots
+- ADD handle active category
 - FIX typo in lost password mail template
 
 ## 1.6 - 2023/11/17